requirements.txt
streamlit
pandas
openpyxl
reportlab
matplotlib
numpy
groq
python-dotenv
scikit-learn
matplotlib
